Integrated the new AJAX API into the existing module
